cleaning some stuffs
git-svn-id: https://openemu.svn.sourceforge.net/svnroot/openemu/OpenEmu@537 a16923d9-94c8-4a39-ae78-11df2d60a2ba

+++ GamePickerController.m
@@ -62,14 +62,14 @@
 	files = [[NSArray arrayWithArray:muteFiles] retain];
 }
 
-- (int)numberOfRowsInTableView:(NSTableView *)tableView
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)tableView
 {
 	return [files count];
 }
 
 - (id)tableView:(NSTableView *)tableView
 objectValueForTableColumn:(NSTableColumn *)tableColumn
-			row:(int)row
+			row:(NSInteger)row
 {
 	return [files objectAtIndex:row];
 }
